+++ lib/aws-operations/aws-config-info-manager.js
@@ -17,7 +17,7 @@ const moment = require('moment')
 const lineByLine = require('n-readlines')
 
 const awsMobileRegions = require('./aws-regions.js').regions
-const awsConfigHelper = require('./aws-config-helper.js')
+const awsConfigProfileReader = require('./aws-config-profile-reader.js')
 const projectInfoManager = require('../project-info-manager.js')
 const awsmobileJSConstant = require('../utils/awsmobilejs-constant.js')
 const nameManager = require('../utils/awsmobilejs-name-manager.js')
@@ -38,7 +38,7 @@ function resolve(silentFlag){
 
 function setProfile(profileName){
     let isSuccessful = true
-    let awsConfig = awsConfigHelper.getSystemConfig(awsInfo.ProfiletName, silentFlag)
+    let awsConfig = awsConfigProfileReader.getSystemConfig(profileName, silentFlag)
     if(awsConfig){
         ensureFolderStructure()
         let awsInfoFilePath = pathManager.getGeneralAWSInfoFilePath()
@@ -65,7 +65,7 @@ function setProfile(profileName){
                 awsInfo.AWSConfigFilePath = path.normalize(path.join(pathManager.getSysProjectAWSConfigDirPath(), nameManager.generateAWSConfigFileName(projectInfo)))
             }
             awsInfo.LastProfileSyncTime = moment().format(awsmobileJSConstant.DateTimeFormatString) 
-            
+
             let jsonString = JSON.stringify(awsInfo, null, '\t')
             fs.writeFileSync(awsInfoFilePath, jsonString, 'utf8')
 
@@ -174,7 +174,7 @@ function resolve_project_awsConfig(projectInfo, silentFlag){
 }
 
 function syncProfile(awsInfo, silentFlag){
-    let awsConfig = awsConfigHelper.getSystemConfig(awsInfo.ProfiletName, silentFlag)
+    let awsConfig = awsConfigProfileReader.getSystemConfig(awsInfo.ProfiletName, silentFlag)
     if(awsConfig){
         let jsonString = JSON.stringify(awsConfig)
         fs.writeFileSync(awsInfo.AWSConfigFilePath, jsonString, 'utf8')
