+++ lib/aws-operations/aws-config-manager.js
@@ -18,7 +18,8 @@ const lineByLine = require('n-readlines')
 const util = require('util')
 
 const awsMobileRegions = require('./aws-regions.js').regions
-const awsConfigFileManager = require('./aws-config-info-manager.js')
+const awsConfigInfoManager = require('./aws-config-info-manager.js')
+const newUserHelper = require('./aws-config-new-user.js')
 const awsmobileJSConstant = require('../utils/awsmobilejs-constant.js')
 const pathManager = require('../utils/awsmobilejs-path-manager.js')
 const nameManager = require('../utils/awsmobilejs-name-manager.js')
@@ -28,11 +29,11 @@ const dfops = require('../utils/directory-file-ops.js')
 function checkAWSConfig(callback){
     let awsDetails = resolveAWSConfig()
 
-    if(awsConfigFileManager.validateAWSConfig(awsDetails.config)){
+    if(awsConfigInfoManager.validateAWSConfig(awsDetails.config)){
         callback(awsDetails)
     }else{
-        if(awsConfigFileManager.isNewUser(awsDetails)){
-            newUserSteps(awsDetails)
+        if(awsConfigInfoManager.isNewUser(awsDetails)){
+            newUserHelper.setupNewUser(awsDetails)
         }else{
             console.log(chalk.red('missing aws account credentials'))
             inquirer.prompt([
@@ -45,7 +46,7 @@ function checkAWSConfig(callback){
             ]).then(function (answers) {
                 if(answers.configureAWS){
                     setConfig(function(awsDetails){
-                        if(awsConfigFileManager.validateAWSConfig(awsDetails.config)){
+                        if(awsConfigInfoManager.validateAWSConfig(awsDetails.config)){
                             callback(awsDetails)
                         }else{
                             console.log(chalk.red('invalid aws account settings'))
@@ -60,13 +61,13 @@ function checkAWSConfig(callback){
 
 function newUserSetup(){
     let awsDetails = resolveAWSConfig(true)
-    if(awsConfigFileManager.validateAWSConfig(awsDetails.config)){
+    if(awsConfigInfoManager.validateAWSConfig(awsDetails.config)){
         console.log('awsmobile-cli already has aws access credentials')
         listAWSConfig(awsDetails)
         enableMobileHubServiceRole()
         promptConfigInstruction()
     }else{
-        newUserSteps(awsDetails)
+        newUserHelper.setupNewUser(awsDetails)
     }
 }
 
@@ -85,111 +86,6 @@ function enableMobileHubServiceRole(){
     console.log('If you see Mobile Hub console, then it is already enabled')
     console.log(chalk.green(awsmobileJSConstant.AWSEnableMobileRoleUrl))
     opn(awsmobileJSConstant.AWSEnableMobileRoleUrl, {wait: false})
-}
-
-function newUserSteps(awsDetails){
-    let awsInfo = awsDetails.info
-    let awsConfig = awsDetails.config
-    
-    console.log('Please follow these steps to setup your awsmobile access')
-    console.log()
-    console.log('Please sign up/in your aws account with Administrator Access:')
-    console.log(chalk.green(awsmobileJSConstant.AWSAmazonUrl))
-    opn(awsmobileJSConstant.AWSAmazonUrl, {wait: false})
-    PressAnyKeyToContinue('Press any key to continue', function(){
-        console.log('Please specify the new IAM user to be created for awsmobile-cli')
-        inquirer.prompt([
-            {
-                type: 'input',
-                name: 'userName',
-                message: "user name: ",
-                default: nameManager.generateIAMUserName()
-            },
-            {
-                type: 'list',
-                name: 'region',
-                message: "region: ",
-                choices: awsMobileRegions,
-                default: 'us-east-1'
-            }
-        ]).then(function (answers) {
-            let deepLinkURL = awsmobileJSConstant.AWSCreateIAMUsersUrl.replace('{userName}', answers.userName).replace('{region}', answers.region)
-            console.log('Please complete the user creation on the aws console')
-            console.log(chalk.green(awsmobileJSConstant.AWSCreateIAMUsersUrl))
-            opn(deepLinkURL, {wait: false})
-            return answers.region
-        }).then(function(region){
-            PressAnyKeyToContinue('Press any key to continue', function(){
-                console.log('Please enter the access key of the newly created user:')
-                inquirer.prompt([
-                    {
-                        type: 'input',
-                        name: 'accessKeyId',
-                        message: "accessKeyId: ",
-                        default: awsmobileJSConstant.DefaultAWSAccessKeyId
-                    },
-                    {
-                        type: 'input',
-                        name: 'secretAccessKey',
-                        message: "secretAccessKey: ",
-                        default: awsmobileJSConstant.DefaultAWSSecretAccessKey
-                    }
-                ]).then(function (answers) {
-                    let awsConfigChanged = false
-                    if(answers.accessKeyId){
-                        let newKeyId = answers.accessKeyId.trim()
-                        if(awsConfig.accessKeyId != newKeyId){
-                            awsConfig.accessKeyId = newKeyId
-                            awsConfigChanged = true
-                        }
-                    }
-                    if(answers.secretAccessKey){
-                        let newKey = answers.secretAccessKey.trim()
-                        if( awsConfig.secretAccessKey != newKey){
-                            awsConfig.secretAccessKey = newKey
-                            awsConfigChanged = true
-                        }  
-                    }
-                    if(region){
-                        let newRegion = region.trim()
-                        if( awsConfig.region != newRegion){
-                            awsConfig.region = newRegion
-                            awsConfigChanged = true
-                        }  
-                    }
-                    if(awsConfigChanged){
-                        let jsonString = JSON.stringify(awsConfig)
-                        fs.writeFileSync(awsInfo.AWSConfigFilePath, jsonString, 'utf8')
-                        awsConfigFileManager.setNoProfileSync(awsInfo)
-                    }
-                    return awsConfig
-                }).then(function(awsConfig){
-                    if(awsConfigFileManager.validateAWSConfig(awsDetails.config)){
-                        console.log()
-                        console.log('Success!')
-                        console.log('You have set the user credential for the awsmobile-cli')
-                        promptConfigInstruction()
-                    }else{
-                        console.log()
-                        console.log('Something went wrong, please try again')
-                    }
-                    process.exit()
-                })
-            })
-        })
-    })
-}
-
-function PressAnyKeyToContinue(message, callback){
-    process.stdin.resume()
-    process.stdin.setEncoding('utf8')
-
-    console.log(message)
-    process.stdin.once('data', function(data){
-        if(callback){
-            callback()
-        }
-    })
 }
 
 function setConfig(callback){
@@ -243,7 +139,7 @@ function setConfig(callback){
         if(awsConfigChanged){
             let jsonString = JSON.stringify(awsConfig)
             fs.writeFileSync(awsInfo.AWSConfigFilePath, jsonString, 'utf8')
-            awsConfigFileManager.setNoProfileSync(awsInfo)
+            awsConfigInfoManager.setNoProfileSync(awsInfo)
         }
 
         if(callback){
@@ -254,7 +150,7 @@ function setConfig(callback){
 
 function setProfile(profileName){
     console.log('Setting awsmobile-cli to use named profile: ' + chalk.blue(profileName))
-    let isSuccessful = awsConfigFileManager.setProfile(profileName) 
+    let isSuccessful = awsConfigInfoManager.setProfile(profileName) 
     if(isSuccessful){
         console.log('Done')
     }else{
@@ -284,7 +180,7 @@ function listAWSConfig(){
 
 function resolveAWSConfig(silentFlag){
     let awsConfig = {}
-    let awsInfo = awsConfigFileManager.resolve(silentFlag)
+    let awsInfo = awsConfigInfoManager.resolve(silentFlag)
     try{
         awsConfig = JSON.parse(fs.readFileSync(awsInfo.AWSConfigFilePath, 'utf8'))
     }catch(e){
